# src/main/resources/swagger-api.yaml
swagger: "2.0"
info:
  version: 1.0.0
  title: Newsreader
  description: Przelotka do newsapi.org
  license:
    name: Apache License 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.txt
host: 127.0.0.1:8080
basePath: /newsreader
schemes:
  - http
tags:
  - name: news
    description: Usługi pobierania wiadomości
paths:
  /news/{country}/{category}:
    get:
      operationId: listArticle
      tags:
       - news
      summary: Pobierz listę wiadomości
      produces:
        - application/json
      parameters:
      - name: country
        in: path
        description: Kod kraju
        required: true
        type: string
      - name: category
        in: path
        description: Kod kategorii (business, entertainment, general, health, science, sports, technology)
        required: true
        type: string
      - name: pageNumber
        in: query
        description: Nr strony (domyślna wartość 1)
        required: false
        type: integer
      - name: pageSize
        in: query
        description: Rozmiar strony (domyślna wartość 10)
        required: false
        type: integer
      - name: searchPhrase
        in: query
        description: Szukany tekst w artykułach
        required: false
        type: string
      responses:
        "200":
          description: Lista wiadomości
          schema:
            $ref: '#/definitions/ArticleList'
        default:
          description: unexpected error
          schema:
            $ref: '#/definitions/Error'
  /categories:
    get:
      operationId: listCategory
      tags:
       - news
      summary: Pobierz listę dostępnych kategorii
      produces:
        - application/json
      responses:
        "200":
          description: Lista katogorii
          schema:
            type: array
            items:
              type: string
        default:
          description: unexpected error
          schema:
            $ref: '#/definitions/Error'
definitions:
  Article:
    required:
      - title
      - date
      - sourceName
      - articleUrl
    properties:
      author:
        type: string
      title:
        type: string
      description:
        type: string
      date:
        type: string
        format: date
      sourceName:
        type: string
      articleUrl:
        type: string
      imageUrl:
        type: string
  ArticleList:
    required:
      - country
      - category
    properties:
      country:
        type: string
      category:
        type: string
      articles:
        type: array
        items:
          $ref: '#/definitions/Article'
      totalCount:
        type: integer
  Error:
    required:
      - errorCode
      - message
    properties:
      errorCode:
        type: string
      message:
        type: string
      fields:
        type: array
        items:
          $ref: '#/definitions/FieldError'
  FieldError:
    required:
      - field
      - message
    properties:
      field:
        type: string
      message:
        type: string